+++ src/components/calcite-pagination/calcite-pagination.tsx
@@ -7,13 +7,16 @@ import {
   Host,
   Prop,
   Method,
-  State,
-  Watch,
 } from "@stencil/core";
 
 import { CSS, TEXT } from "./resources";
 
 const maxPagesDisplayed = 5;
+export interface CalcitePaginationDetail {
+  start: number;
+  total: number;
+  num: number;
+}
 
 @Component({
   tag: "calcite-pagination",
@@ -26,49 +29,30 @@ export class CalcitePagination {
   //  Public Properties
   //
   //--------------------------------------------------------------------------
+  /** number of items per page */
+  @Prop() num = 20;
 
-  /** Change between foreground colors or background colors for container background */
-  @Prop({ reflect: true }) backgroundStyle:
-    | "backgroundColor"
-    | "foregroundColor" = "foregroundColor";
+  /** index of item that should begin the page */
+  @Prop() start = 1;
 
-  /** starting selected index */
-  @Prop({ reflect: true }) num = 1;
-  @Watch("num") numWatchHandler(newValue) {
-    this.selectedIndex = newValue;
-  }
-
-  /** starting number of the pagination */
-  @Prop({ reflect: true }) start = 1;
+  /** total number of items */
+  @Prop() total = 0;
 
   /** title of the next button */
-  @Prop({ reflect: true }) textLabelNext: string = TEXT.nextLabel;
+  @Prop() textLabelNext: string = TEXT.nextLabel;
 
   /** title of the previous button */
-  @Prop({ reflect: true }) textLabelPrevious: string = TEXT.previousLabel;
+  @Prop() textLabelPrevious: string = TEXT.previousLabel;
 
   /** specify the theme of accordion, defaults to light */
   @Prop({ reflect: true }) theme: "light" | "dark";
-
-  /** ending number of the pagination */
-  @Prop({ reflect: true }) total = 2;
 
   // --------------------------------------------------------------------------
   //
   //  Private Properties
   //
   // --------------------------------------------------------------------------
-
   @Element() el: HTMLElement;
-
-  @State() selectedIndex = this.num;
-  @Watch("selectedIndex") selectedIndexWatchHandler() {
-    this.calcitePaginationUpdate.emit({
-      start: this.start,
-      total: this.total,
-      num: this.selectedIndex,
-    });
-  }
 
   //--------------------------------------------------------------------------
   //
@@ -79,7 +63,7 @@ export class CalcitePagination {
   /** Emitted whenever the selected page changes.
    * @event calcitePaginationUpdate
    */
-  @Event() calcitePaginationUpdate: EventEmitter;
+  @Event() calcitePaginationUpdate: EventEmitter<CalcitePaginationDetail>;
 
   // --------------------------------------------------------------------------
   //
@@ -87,25 +71,14 @@ export class CalcitePagination {
   //
   // --------------------------------------------------------------------------
 
-  /** When called, selected page will increment by 1.
-   */
-  @Method()
-  async nextPage(): Promise<void> {
-    this.selectedIndex = Math.min(this.total, this.selectedIndex + 1);
+  /** Go to the next page of results */
+  @Method() async nextPage(): Promise<void> {
+    this.start = Math.min(this.getLastStart(), this.start + this.num);
   }
 
-  /** When called, selected page will decrement by 1.
-   */
-  @Method()
-  async previousPage(): Promise<void> {
-    this.selectedIndex = Math.max(this.start, this.selectedIndex - 1);
-  }
-
-  /** Set selected page to a specific page number. Will not go below start or above total.
-   */
-  @Method()
-  async setPage(num: number): Promise<void> {
-    this.selectedIndex = Math.max(this.start, Math.min(this.total, num));
+  /** Go to the previous page of results */
+  @Method() async previousPage(): Promise<void> {
+    this.start = Math.max(1, this.start - this.num);
   }
 
   // --------------------------------------------------------------------------
@@ -114,20 +87,37 @@ export class CalcitePagination {
   //
   // --------------------------------------------------------------------------
 
-  previousClicked = (): void => {
-    this.previousPage();
-  };
-
-  nextClicked = (): void => {
-    this.nextPage();
-  };
-
-  showLeftEllipsis() {
-    return this.selectedIndex - this.start > 3;
+  private getLastStart(): number {
+    const { total, num } = this;
+    const lastStart =
+      total % num === 0 ? total - num : Math.floor(total / num) * num;
+    return lastStart + 1;
   }
 
-  showRightEllipsis() {
-    return this.total - this.selectedIndex > 3;
+  private previousClicked = (): void => {
+    this.previousPage().then();
+    this.emitUpdate();
+  };
+
+  private nextClicked = (): void => {
+    this.nextPage();
+    this.emitUpdate();
+  };
+
+  private showLeftEllipsis() {
+    return Math.floor(this.start / this.num) > 3;
+  }
+
+  private showRightEllipsis() {
+    return (this.total - this.start) / this.num > 3;
+  }
+
+  private emitUpdate() {
+    this.calcitePaginationUpdate.emit({
+      start: this.start,
+      total: this.total,
+      num: this.num,
+    });
   }
 
   //--------------------------------------------------------------------------
@@ -137,52 +127,60 @@ export class CalcitePagination {
   //--------------------------------------------------------------------------
 
   renderPages() {
-    let pages = [];
-    let currentNum;
+    let lastStart = this.getLastStart();
     let end;
+    let nextStart;
 
-    if (this.total <= maxPagesDisplayed) {
-      currentNum = this.start + 1;
-      end = this.total - 1;
+    // if we don't need ellipses render the whole set
+    if (this.total / this.num <= maxPagesDisplayed) {
+      nextStart = 1 + this.num;
+      end = lastStart - this.num;
     } else {
-      if (this.selectedIndex < maxPagesDisplayed) {
-        currentNum = this.start + 1;
-        end = this.start + 4;
+      // if we're within max pages of page 1
+      if (this.start / this.num < maxPagesDisplayed - 1) {
+        nextStart = 1 + this.num;
+        end = 1 + 4 * this.num;
       } else {
-        if (this.selectedIndex + 3 >= this.total) {
-          currentNum = this.total - 4;
-          end = this.total - 1;
+        // if we're within max pages of last page
+        if (this.start + 3 * this.num >= this.total) {
+          nextStart = lastStart - 4 * this.num;
+          end = lastStart - this.num;
         } else {
-          currentNum = this.selectedIndex - 1;
-          end = this.selectedIndex + 1;
+          nextStart = this.start - this.num;
+          end = this.start + this.num;
         }
       }
     }
 
-    while (currentNum <= end) {
-      pages.push(currentNum);
-      currentNum++;
+    const pages = [];
+    while (nextStart <= end) {
+      pages.push(nextStart);
+      nextStart = nextStart + this.num;
     }
 
     return pages.map((page) => this.renderPage(page));
   }
 
-  renderPage(num) {
+  renderPage(start: number) {
+    const page = Math.floor(start / this.num) + 1;
     return (
-      <a
-        tabIndex={0}
-        class={{ [CSS.page]: true, [CSS.selected]: num === this.selectedIndex }}
+      <button
+        class={{
+          [CSS.page]: true,
+          [CSS.selected]: start === this.start,
+        }}
         onClick={() => {
-          this.selectedIndex = num;
+          this.start = start;
+          this.emitUpdate();
         }}
       >
-        {num}
-      </a>
+        {page}
+      </button>
     );
   }
 
   renderLeftEllipsis() {
-    if (this.total > maxPagesDisplayed && this.showLeftEllipsis()) {
+    if (this.total / this.num > maxPagesDisplayed && this.showLeftEllipsis()) {
       return (
         <span class={`${CSS.ellipsis} ${CSS.ellipsisStart}`}>
           <calcite-icon scale="s" icon="ellipsis" />
@@ -192,7 +190,7 @@ export class CalcitePagination {
   }
 
   renderRightEllipsis() {
-    if (this.total > maxPagesDisplayed && this.showRightEllipsis()) {
+    if (this.total / this.num > maxPagesDisplayed && this.showRightEllipsis()) {
       return (
         <span class={`${CSS.ellipsis} ${CSS.ellipsisEnd}`}>
           <calcite-icon scale="s" icon="ellipsis" />
@@ -202,35 +200,36 @@ export class CalcitePagination {
   }
 
   render() {
+    const { total, num, start } = this;
     return (
-      <Host class={this.backgroundStyle}>
-        <a
+      <Host>
+        <button
           class={{
             [CSS.previous]: true,
-            [CSS.disabled]: this.selectedIndex <= 1,
+            [CSS.disabled]: start < num,
           }}
-          tabIndex={0}
-          title={this.textLabelPrevious}
+          aria-label={this.textLabelPrevious}
           onClick={this.previousClicked}
+          disabled={start < num}
         >
           <calcite-icon scale="s" icon="chevronLeft" />
-        </a>
-        {this.renderPage(this.start)}
+        </button>
+        {this.renderPage(1)}
         {this.renderLeftEllipsis()}
         {this.renderPages()}
         {this.renderRightEllipsis()}
-        {this.renderPage(this.total)}
-        <a
+        {this.renderPage(this.getLastStart())}
+        <button
           class={{
             [CSS.next]: true,
-            [CSS.disabled]: this.selectedIndex >= this.total,
+            [CSS.disabled]: start + num >= total,
           }}
-          tabIndex={0}
-          title={this.textLabelNext}
+          aria-label={this.textLabelNext}
           onClick={this.nextClicked}
+          disabled={start + num >= total}
         >
           <calcite-icon scale="s" icon="chevronRight" />
-        </a>
+        </button>
       </Host>
     );
   }